2024 Theme: Move course grid block to theme
It is a presentation component which is tightly coupled to the theme.

See https://github.com/WordPress/Learn/issues/2573
