+++ wp-content/themes/pub/wporg-learn-2024/functions.php
@@ -5,6 +5,7 @@ namespace WordPressdotorg\Theme\Learn_2024;
 use function WPOrg_Learn\Sensei\{get_my_courses_page_url};
 
 // Block files
+require_once __DIR__ . '/src/course-grid/index.php';
 require_once __DIR__ . '/src/learning-pathway-cards/index.php';
 require_once __DIR__ . '/src/learning-pathway-header/index.php';
 require_once __DIR__ . '/src/search-results-context/index.php';
